+++ packages/hooks/use-image/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nextui-org/use-image
 
+## 0.0.0-dev-v2-20230326032643
+
+### Patch Changes
+
+- Calculated styled added to theme components, icon button improved
+- Updated dependencies
+  - @nextui-org/use-safe-layout-effect@0.0.0-dev-v2-20230326032643
+
 ## 0.0.0-dev-v2-20230326015124
 
 ### Patch Changes